requirements.txt
streamlit~=1.42.2
torch~=2.6.0
torchvision~=0.21.0
matplotlib~=3.10.0
seaborn~=0.13.2
numpy~=2.2.3
pillow~=11.1.0
requests~=2.32.3
setuptools~=68.2.0
Jinja2~=3.1.5
wheel~=0.41.2